test: Re-organize tests a bit
Move some tests to more suitable sub-directories.

Rename files to all start with test- so they don't conflict with the C
library source file names.

Restructure test/meson.build to reduce number of special cases.
