+++ test/test-hosted-exit.c
