+++ test/test-math/test-rounding-mode-sub.c
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: BSD-3-Clause
  *
- * Copyright © 2020 Keith Packard
+ * Copyright © 2021 Keith Packard
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -33,13 +33,43 @@
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "try-ilp32.h"
+#include <stdio.h>
+#include <math.h>
+#include <fenv.h>
+#include "test-rounding-mode.h"
 
-struct strdesc
-get_str(char *sample)
+double
+div_d(double a, double b)
 {
-    struct strdesc x;
-    x.base = sample;
-    x.len = strlen(sample);
-    return x;
+    return a / b;
+}
+
+double
+mul_d(double a, double b)
+{
+    return a * b;
+}
+
+double
+sub_d(double a, double b)
+{
+    return a - b;
+}
+
+float
+div_f(float a, float b)
+{
+    return a / b;
+}
+
+float
+mul_f(float a, float b)
+{
+    return a * b;
+}
+
+float
+sub_f(float a, float b)
+{
+    return a - b;
 }
