+++ test/test-regex.c
