+++ test/test-stdlib/test-atexit.c
