+++ test/test-stdlib/test-on_exit.c
