+++ test/test-stdlib/test-rand.c
