+++ test/test-string/test-ffs.c
